ai-lab2/requirements.txt
langchain 
langchain_groq
langgraph
python-dotenv
streamlit


